fix(review): normalize 'self' subject on writes; doc + test hardening
Whole-PR xhigh review follow-ups:
- subject:"self" now normalizes to NULL on both write paths (record +
  update), closing the write/read sentinel asymmetry; update_claim's
  subject:"self" doubles as the reclassify-to-Steve affordance, and the
  set-only/no-clear limitation is now documented like update_profile's
- ADR-0004 §5 + entity comment reworded: general is the coached
  destination; out-of-vocabulary categories are schema-rejected with a
  retryable error, not silently quarantined (doc now matches code)
- #[serde(default)] on record_intake_answers claims: absent key reaches
  the actionable domain error
- confidence-vocabulary sync test ties prompt prose to ClaimConfidence
- shared assert_scripts_in_order test helper replaces two divergent
  ordered-token idioms (absolute-find could not pin repeated tokens)
- CategoryCoverage/IntakeState counts are usize (drops try_from ceremony)

+++ healthie-mcp/src/handler.rs
@@ -47,8 +47,8 @@ pub struct HealthieMcp {
 #[derive(Serialize)]
 struct IntakeState {
     coverage: Vec<claim::CategoryCoverage>,
-    total_claims: u64,
-    total_unknowns: u64,
+    total_claims: usize,
+    total_unknowns: usize,
 }
 
 #[tool_router]
